+++ voting-system/package.json
@@ -11,16 +11,12 @@
     "object-hash": "^2.0.0",
     "react": "^16.10.1",
     "react-bootstrap": "^1.0.0-beta.14",
-    "react-circular-progressbar": "^2.0.2",
     "react-csv-reader": "^1.3.1",
     "react-dom": "^16.10.1",
-    "react-icons": "^3.7.0",
     "react-loading": "^2.0.3",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.1.2",
-    "react-top-loading-bar": "^1.2.0",
-    "semantic-ui-calendar-react": "^0.15.3",
-    "semantic-ui-react": "^0.88.1",
+    "react-slick": "^0.25.2",
     "typescript": "^3.6.3"
   },
   "scripts": {
