+++ voting-system/public/favicon.ico
