+++ voting-system/public/index.html
@@ -37,10 +37,6 @@
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" />
     <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons" />
     <title>React App</title>
-
-    <!-- semantic -->
-    <link rel="stylesheet" href="//cdn.jsdelivr.net/npm/semantic-ui@2.4.2/dist/semantic.min.css" />
-
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
