+++ voting-system/public/logo192.png
