+++ voting-system/src/App.jsx
@@ -3,8 +3,8 @@ import { BrowserRouter as Router, Route, Redirect, Switch } from "react-router-d
 import { Container, Row } from "react-bootstrap";
 import MUIContainer from '@material-ui/core/Container';
 import './App.css';
-import Login from './Views/Login';
-import Home from './Views/Home';
+import Login from './Views/Login.jsx';
+import Home from './Views/Home.jsx';
 import RouteObjects from './routes.js';
 import ReactLoading from 'react-loading';
 import { firebase } from './Firebase.jsx';
@@ -15,7 +15,6 @@ class App extends Component {
     super(props);
     this.state = ({
       user: null,
-      userDetails: null,
       fetchingUser: true
     });
     this.authListener = this.authListener.bind(this);
@@ -40,8 +39,9 @@ class App extends Component {
           if (!doc.exists) {
             this.setState({ user: null, fetchingUser: false });
           } else {
-            console.log(doc.data());
-            this.setState({ user: doc.data(), userDetails: user, fetchingUser: false });
+            var theUser = doc.data();
+            theUser.id = doc.id;
+            this.setState({ user: theUser, fetchingUser: false });
           }
         })
         .catch(err => {
@@ -49,7 +49,6 @@ class App extends Component {
         })
       }
       else {
-        console.log("Error getting user")
         this.setState({ user: null, fetchingUser: false });
       }
     });
@@ -60,7 +59,7 @@ class App extends Component {
       return (
         <Route
           path={route.path}
-          render={props => (<route.component user={this.state.user} userDetails={this.state.userDetails} {...props}/>)}
+          render={props => (<route.component {...props} user={this.state.user}/>)}
           key={key}
         />
       );
@@ -68,9 +67,17 @@ class App extends Component {
   };
 
   render() {
-    // if havn't fetch user display loading bar
+    // If user hasn't been fetched yet
     if(this.state.fetchingUser){
-      return <ReactLoading type="bubbles" color="blue" height={667} width={375} />
+      return (
+        <div className="App container">
+          <MUIContainer component="main" maxWidth="xs">
+            <Container fluid >
+              <ReactLoading type="bubbles" color="blue" height={667} width={375} />
+            </Container>
+          </MUIContainer>
+        </div>
+      )
     }
     // Router logic
     var renderPath = "";
@@ -99,7 +106,7 @@ class App extends Component {
             <Router>
               <Switch>
                 { this.state.user && this.getRoutes(RouteObjects) }
-                <Route path={renderPath} render={props => <RenderPage user={this.state.user} userDetails={this.state.userDetails} {...props} />} />
+                <Route path={renderPath} render={props => <RenderPage {...props} user={this.state.user} />} />
                 <Redirect from="/" to={redirectPath} />
               </Switch>
             </Router>
