+++ voting-system/src/Views/Login.jsx
@@ -88,7 +88,6 @@ MySnackbarContentWrapper.propTypes = {
   variant: PropTypes.oneOf(['error', 'info', 'success', 'warning']).isRequired,
 };
 
-// login class
 class Login extends Component {
   _isMounted = false;
   constructor(props) {
@@ -112,7 +111,6 @@ class Login extends Component {
     e.preventDefault();
     firebase.auth().signInWithEmailAndPassword(this.state.email, this.state.password)
     .then(firebaseUser => {
-      console.log(firebaseUser)
       console.log("Signed in!");
     })
     .catch(error => {
