+++ voting-system/src/index.js
